react native project
this project sends an email to me in a form of a json file.

I tried connecting it to a mongo db but I run out of time.

There is one error that I wasn't able to figure out but shouldn't cause any problem to the code.

The call phone isn't responsive.

The code takes longer time to run because I was try to connect to to mongodb which didn't work out.